fix: Import project with Unity 5.6.0b7 & Google VR SDK V1.20
-GVR Raycaster/Pointer/Reticle now working properly

Note: Cardboard VR no longer works in Editor/Game View Mode, as per the console message from Google. It is only supported when deployed to a supported Android device. (This will not hinder development because I can disable VR mode and test the scene full screen, then deploy to a device when ready for further testing)

Resolves: dcb26ec